oups

+++ trunk/fflas-ffpack/field/modular-uint64.h
@@ -106,13 +106,13 @@ namespace FFPACK
 			return x;
 		}
 
-		Element &init (Element &x, const uint64_t &y ) const
+		Element &init (Element &x, const int64_t &y ) const
 		{
 			x = Element(y %  (_modulus));
 			return x;
 		}
 
-		Element &init (Element &x, const uint32_t &y ) const
+		Element &init (Element &x, const uint64_t &y ) const
 		{
 			x = Element(y %  (_modulus));
 			return x;
